rename classes.py > medication.py
